Difficulty tweaks and some changes
I tweaked the difficulty settings so the higher the difficulty, the
faster the enemies, sooner the spawning, and shorter spawning times.

Since I could not figure our hot to implement a arbitrary shot limit
smoothly, I have removed it altogether. To compensate it, the player's
bullets shoot faster to keep up with the pace with the fast enemies.